+++ ProcessIndieAuth.config.php
@@ -19,6 +19,13 @@ $config = [
         'value' => 1209600,
         'required' => true,
     ],
+    'only_advertise_metadata' => [
+        'type' => 'checkbox',
+        'label' => 'Only advertise the IndieAuth Server Metadata endpoint',
+        'description' => 'Optional. Enable this if you want to remove the backwards-compatible link elements for  rel="authorization_endpoint" and rel="token_endpoint"',
+        'value' => 0,
+        'collapsed' => Inputfield::collapsedYes,
+    ],
     'auto_revoke' => [
         'type' => 'checkbox',
         'label' => 'Automatically remove tokens after expiration',
@@ -28,6 +35,10 @@ $config = [
     ],
     'token_secret' => [
         'type' => 'hidden',
+    ],
+    'schema_version' => [
+        'type' => 'hidden',
+        'value' => 0,
     ],
 ];