+++ CHANGELOG.md
@@ -4,6 +4,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.2.3] - 2025-01-26
+### Fixed
+- Adding profile information to token responses was not working correctly [#6](https://github.com/gRegorLove/ProcessWire-IndieAuth/issues/6)
+- Install attempts to create the introspection-endpoint page [#7](https://github.com/gRegorLove/ProcessWire-IndieAuth/issues/7)
+
+### Changed
+- Admin: improve list of granted access in admin [#8](https://github.com/gRegorLove/ProcessWire-IndieAuth/issues/8)
+
+### Added
+- Add ability to disable backwards-compatible link-rels [#9](https://github.com/gRegorLove/ProcessWire-IndieAuth/issues/9)
+- Database schema version and automatic update process
+
 ## [0.2.2] - 2024-07-12
 ### Added
 - Introspection Endpoint with HTTP Basic Authentication support